requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
bardapi==0.1.38
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
deep-translator==1.11.4
exceptiongroup==1.1.3
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
jeepney==0.8.0
langdetect==1.0.9
lz4==4.3.2
multidict==6.0.4
numpy==1.26.0
openai==0.28.1
pandas==2.1.1
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sec-api==1.0.17
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
